Move __init__.py contents to endpoints.py. Use FlaskForm instead of Flask due to deprecation.
