+++ app/mod_main/forms.py
@@ -1,12 +1,12 @@
 # app/mod_main/forms.py
 
 
-from flask_wtf import Form
+from flask_wtf import FlaskForm
 from wtforms import TextField
 from wtforms.validators import DataRequired, Email, length
 
 
-class SignUpForm(Form):
+class SignUpForm(FlaskForm):
     email = TextField(
         'Enter your email address',
         validators=[DataRequired(), Email(), length(min=3)])
